requirements.txt
streamlit
openai-whisper
librosa
soundfile
langchain-openai
langgraph
langchain-tavily
python-dotenv
langchain-core